requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
facenet-pytorch==2.5.3
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
fonttools==4.50.0
fsspec==2024.3.1
grad-cam==1.5.0
gradio==4.22.0
gradio_client==0.13.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.9.15
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.18.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.36.3
sympy==1.12
threadpoolctl==3.4.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
ttach==0.0.3
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
Werkzeug==3.0.1
wsproto==1.2.0
